+++ convertors/html_convertor/query.py
@@ -3,27 +3,61 @@
 import logging
 from lxml import etree
 import lxml.html
-from lxml.etree import XPathError
+from lxml.etree import XPathError, XPathEvalError
 #from lxml.html.clean import clean_html
 from html_table_parser import HTMLTableParser
 #from html_table_extractor.extractor import Extractor
+from utils import match
+from utils.property_str import pstr
+from html2text import HTML2Text
 
 import utils.logging
 logger = logging.getLogger(utils.logging.getLoggerName(__name__))
 
 XPATH_FIELD = "xpath"
+ATTR_NAME = "attribute-name"
 COLUMN_NUM = "column-num"
 ROW_NUM = "row-num"
 PROCESS_IGNORE_COLS = "ignore-cols"
 PROCESS_REMOVE_HEADER_ROW = 'remove-header-row'
 PROCESS_REMOVE_ROWS_IF_EMPTY = "remove-rows-if-empty"
 PROCESS_SPLIT_TYPE = "split-type"
-
-# def _strip_element_list(ele_list):
-
-#     for ele in ele_list:
-#         etree.strip_elements(ele, "*")
-#     return
+HTML2TEXT_OPTIONS = "html2text-options"
+HTML2TEXT_ALLOWLIST = [
+    "unicode_snob",
+    "escape_snob",
+    "links_each_paragraph",
+    "body_width",
+    "skip_internal_links",
+    "inline_links",
+    "protect_links",
+    "google_list_indent",
+    "ignore_links",
+    "ignore_mailto_links",
+    "ignore_images",
+    "images_as_html",
+    "images_to_alt",
+    "images_with_size",
+    "ignore_emphasis",
+    "bypass_tables",
+    "ignore_tables",
+    "google_doc",
+    "ul_item_mark",
+    "emphasis_mark", 
+    "strong_mark",
+    "single_line_break",
+    "use_automatic_links",
+    "hide_strikethrough",
+    "mark_code",
+    "wrap_list_items",
+    "wrap_links",
+    "wrap_tables",
+    "pad_tables",
+    "default_image_alt",
+    "open_quote", 
+    "close_quote", 
+    "include_sup_sub"
+]
 
 def get_document(document_str, mapping):
 
@@ -42,12 +76,16 @@ def get_document(document_str, mapping):
                 continue
             
             for ele in selection:
-                if tags_to_strip := processor.get("strip-tags", None):
-                    etree.strip_tags(ele, tags_to_strip)
-                if elements_to_strip := processor.get("strip-elements", None):
-                    etree.strip_elements(ele, elements_to_strip)
+                # Order here matters, do attributes first because that is irresptive of tags or elements
                 if attributes_to_strip := processor.get("strip-attributes", None):
                     etree.strip_attributes(ele, attributes_to_strip)
+                # Do elements before tags because we may want to strip all tags and so if we did tags first we couldn't strip the elements
+                if elements_to_strip := processor.get("strip-elements", None):
+                    etree.strip_elements(ele, elements_to_strip)
+                if tags_to_strip := processor.get("strip-tags", None):
+                    etree.strip_tags(ele, tags_to_strip)
+                
+                
 
     # This is a hack.  For reasons unknown XPath queries start failing for no obvious reason
     # once strip_* methods are called.  So we reload the processed XML from a string.  Lame.
@@ -57,15 +95,6 @@ def get_document(document_str, mapping):
 
     return document
 
-    # try:
-    #     h1_list = document.xpath("//h1")
-    #     _strip_element_list(h1_list)
-    # except XPathError:
-    #     logger.warning(f"XPath query to find headers caused an error")
-
-    # return document
-    # Note, can't do much pre-processing of the XML here as the scheme might be referencing something that we might change
-    #return lxml.html.document_fromstring(document_str)
 
 def get_document_section(document, query_cfg):
 
@@ -120,12 +149,65 @@ def get_document_value(document, query_cfg) -> str:
     if len(value_list) > 1:
         logger.warning(f"Retreived {len(value_list)} document values when 1 was expected.  Using first.")
         
+    location = ""
+
     if not isinstance(value_list[0], str):
         logger.warning(f"Expected results 'str', got '{type(value_list[0])}', for query '{query_cfg[XPATH_FIELD]}'")
-
+    else:
+        # It's a string, so we can't extract a location from it.  Try to strip away the xpath path that makes the result a string to get to the
+        # element that contains the string.
+        query = query_cfg[XPATH_FIELD]
+        if query.endswith("//text()"):
+            query = query[:-8]
+        if query.endswith("/text()"):
+            query = query[:-7]
+        elif "/@" in query:
+            query = query[0:query.rfind("/@")]
+        
+        try:
+            value_element = document.xpath(query)
+        except (XPathError, XPathEvalError):
+            logger.warning(f"XPath query '{query}' caused an error")
+            value_list = []
+        if len(value_element) == 1:
+            if hasattr(value_element[0], "xpath"):
+                location = document.getroottree().getpath(value_element[0])
+        
     # Despite checking this is a str, it could be a _ElementUnicodeResult, which doesn't serialise to JSON nicely.  So explicitly convert.
 
-    return str(value_list[0])
+    if location:
+        return pstr(str(value_list[0]), properties = {"location":location})
+    
+    logger.info(f"Returning value '{str(value_list[0])}' without location")
+    return pstr(str(value_list[0]))
+
+def get_element_text(document, query_cfg) -> str:
+
+    if (element := _get_element(document, query_cfg)) is None:
+        return None
+    
+    text = element.text_content()
+
+    # The output from lxml is usually not a string, so explicitly convert
+    return pstr(str(text), properties = {"location": element.getroottree().getpath(element)})
+
+
+def get_element_attribute(document, query_cfg) -> str:
+
+    if (element := _get_element(document, query_cfg)) is None:
+        return None
+    
+    if not query_key_defined(query_cfg, ATTR_NAME):
+        logger.warning(f"'{ATTR_NAME}' not specified in query configuration")
+
+    attribute = element.get(query_cfg[ATTR_NAME], None)
+
+    if attribute is None:
+        logger.warning(f"Attribute '{query_cfg[ATTR_NAME]}' not found in element")
+
+    # The output from lxml is usually not a string, so explicitly convert
+    return pstr(str(attribute), properties = {"location": element.getroottree().getpath(element)})
+    
 
 def _remove_rows_if_empty(proc_def, table_data):
 
@@ -152,6 +234,24 @@ def _remove_rows_if_empty(proc_def, table_data):
 
     return new_table_data
 
+def get_table_xpaths(table_element):
+    row_xpaths = []
+    roottree = table_element.getroottree()
+    rows = table_element.xpath('.//tr')
+    
+    for row_index, row in enumerate(rows):
+        col_xpaths = []
+        cells = row.xpath('.//td | .//th')
+        for cell_index, cell in enumerate(cells):
+            # Since the cell element could be a td or th, we must make sure we get the correct xpath
+            #cell_xpath = f"{table_element.getroottree().getpath(table_element)}//tr[{row_index + 1}]//{cell.tag}[{cell_index + 1}]"
+            cell_xpath = roottree.getpath(cell)    # This is dangerous though as pre-processing may make the xpath here invalid when applied to the original document
+            col_xpaths.append(cell_xpath)
+        
+        row_xpaths.append(col_xpaths)
+    
+    return row_xpaths
+
 def get_document_row_table(document, query_cfg):
 
     if not query_key_defined(query_cfg, XPATH_FIELD):
@@ -171,6 +271,9 @@ def get_document_row_table(document, query_cfg):
         logger.warning(f"XPath query '{query_cfg[XPATH_FIELD]}' returned more than 1 result, only using first")
 
     table_ele = table_list[0]
+
+    # Get the XPath to each table element
+    table_xpaths = get_table_xpaths(table_ele)
 
     # For future reference:
     #  'strip_tags' will remove the element tag and attributes but not the content
@@ -233,6 +336,29 @@ def get_document_row_table(document, query_cfg):
     # Return list of lists.  There is only 1 table so only the first is returned
     table_output = p.tables[0]
 
+    # Combine the XPath to each table cell wit the value in each table cell
+    table_values = table_output
+    if len(table_values) != len(table_xpaths):
+        # We'll cope with this by using empty XPath vlaues
+        logger.warning(f"Table values '{len(table_values)}' did not match the table xpaths '{len(table_xpaths)}'")
+    for row_index, row in enumerate(table_values):
+        if row_index >= len(table_xpaths):
+            # Make entire row empty XPath values
+            logger.warning(f"Table value row '{row_index}' did not have a corresponding table xpath")
+            row_xpath = [""]*len(row)
+        else:
+            row_xpath = table_xpaths[row_index]
+        for col_index, value in enumerate(row):
+            if col_index >= len(row_xpath):
+                # Make column XPath entry empty.  Remember, this row could contain real XPath values, just not enough of them
+                logger.warning(f"Table value '{value}' did not have a corresponding table xpath")
+                col_xpath = ""
+            else:
+                col_xpath = row_xpath[col_index]
+            
+            #table_output[row_index][col_index] = {"location":col_xpath, "value":value}
+            table_output[row_index][col_index] = pstr(value, properties={"location":col_xpath})
+
     if query_cfg.get(PROCESS_REMOVE_HEADER_ROW, False):
         table_output = table_output[1:]
 
@@ -293,7 +419,7 @@ def get_constrained_table_entry(row, query_cfg, current_row_index, current_col_i
 
     return None
 
-def query_key_defined(query_cfg, query_cfg_key):
+def query_key_defined(query_cfg, query_cfg_key, optional = False):
 
     if query_cfg is None:
         logger.warning(f"Query configuration is not specified")
@@ -304,10 +430,31 @@ def query_key_defined(query_cfg, query_cfg_key):
         return False
 
     if query_cfg_key not in query_cfg:
-        logger.warning(f"Query configuration '{COLUMN_NUM}' was not present in '{query_cfg}'")
+        if not optional:
+            logger.warning(f"Query configuration '{query_cfg_key}' was not present in '{query_cfg}'")
         return False
     
     return True
+
+def _get_element(document, query_cfg) -> str:
+
+    if not query_key_defined(query_cfg, XPATH_FIELD):
+        return None
+
+    try:
+        value_list = document.xpath(query_cfg[XPATH_FIELD])
+    except XPathError:
+        logger.warning(f"XPath query '{query_cfg[XPATH_FIELD]}' caused an error")
+        value_list = []
+
+    logger.debug(f"Query '{query_cfg[XPATH_FIELD]}' return {len(value_list)} elements")
+
+    if len(value_list) == 0:
+        return None
+    if len(value_list) > 1:
+        logger.warning(f"Retreived {len(value_list)} document values when 1 was expected.  Using first.")
+
+    return value_list[0]
 
 def does_col_index_match(query_cfg, index):
 
@@ -339,23 +486,33 @@ def set_table_entry(row, query_cfg, value):
 
     return
 
-# Expecting the xpath to return a list of text nodes,whcih get concatenated
+# Expecting the xpath to return a list of nodes, which get text exrtacted and concatenated
 def get_text_section(document, query_cfg):
 
     if not query_key_defined(query_cfg, XPATH_FIELD):
         return None
 
     try: 
-        paragraphs = document.xpath(query_cfg[XPATH_FIELD])
+        nodes = document.xpath(query_cfg[XPATH_FIELD])
     except XPathError:
         logger.warning(f"XPath query '{query_cfg[XPATH_FIELD]}' caused an error")
-        paragraphs = []
+        nodes = []
 
-    output = ""
-    for paragraph in paragraphs:
-        output += (str(paragraph) + "/n")
+    input = ""
+    for node in nodes:
+        input += lxml.html.tostring(node, encoding='unicode')
 
-    return output
+    h2t = HTML2Text()
+    if query_key_defined(query_cfg, HTML2TEXT_OPTIONS, optional=True) and isinstance(query_cfg[HTML2TEXT_OPTIONS], dict):
+        for option, value in query_cfg[HTML2TEXT_OPTIONS].items():
+            if isinstance(option, str) and option.lower() in HTML2TEXT_ALLOWLIST:
+                setattr(h2t, option.lower(), value)
+            else:
+                logger.warning(f"html2text option '{option.lower()}' not in allowed list")
+    
+    output = h2t.handle(str(input))
+
+    return pstr(output, properties={"location":query_cfg[XPATH_FIELD]})
 
 html_dispatch_table = {
     "html-ul":get_document_list_items,
@@ -365,5 +522,7 @@ html_dispatch_table = {
     "html-split-table":get_split_table,
     "html-table-transpose":get_transposed_table,
     "html-text":get_document_value,
+    "html-element-text":get_element_text,
+    "html-element-attribute": get_element_attribute,
     "html-text-section":get_text_section
 }